redcoding asset handler, holder, tracker, memory asset
