+++ Gems/ScriptCanvas/Code/Editor/View/Windows/Tools/UpgradeTool/Model.cpp
@@ -8,7 +8,6 @@
 
 #include <AzCore/Asset/AssetManagerBus.h>
 #include <CryCommon/CrySystemBus.h>
-#include <Editor/Assets/ScriptCanvasAssetHelpers.h>
 #include <Editor/View/Windows/Tools/UpgradeTool/Model.h>
 #include <IConsole.h>
 #include <ISystem.h>
