+++ Gems/ScriptCanvas/Code/scriptcanvasgem_editor_files.cmake
@@ -15,12 +15,8 @@ set(FILES
     Editor/SystemComponent.h
     Editor/SystemComponent.cpp
     Editor/QtMetaTypes.h
-    Editor/Assets/ScriptCanvasAssetTracker.cpp
-    Editor/Assets/ScriptCanvasAssetTracker.h
-    Editor/Assets/ScriptCanvasAssetTrackerBus.h
     Editor/Assets/ScriptCanvasAssetHelpers.h
     Editor/Assets/ScriptCanvasAssetHelpers.cpp
-    Editor/Assets/ScriptCanvasAssetTrackerDefinitions.h
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasBaseAssetData.h
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasBaseAssetData.cpp
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasAsset.h
@@ -28,13 +24,7 @@ set(FILES
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasAssetBus.h
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasAssetTypes.h
     Editor/Include/ScriptCanvas/Assets/ScriptCanvasFileHandling.h
-    Editor/Include/ScriptCanvas/Assets/ScriptCanvasAssetHandler.h
     Editor/Assets/ScriptCanvasFileHandling.cpp
-    Editor/Assets/ScriptCanvasAssetHandler.cpp
-    Editor/Assets/ScriptCanvasAssetHolder.h
-    Editor/Assets/ScriptCanvasAssetHolder.cpp
-    Editor/Assets/ScriptCanvasMemoryAsset.h
-    Editor/Assets/ScriptCanvasMemoryAsset.cpp
     Editor/Assets/ScriptCanvasUndoHelper.h
     Editor/Assets/ScriptCanvasUndoHelper.cpp
     Editor/Include/ScriptCanvas/Bus/RequestBus.h
@@ -56,6 +46,8 @@ set(FILES
     Editor/Include/ScriptCanvas/Components/EditorGraphVariableManagerComponent.h
     Editor/Components/EditorGraphVariableManagerComponent.cpp
     Editor/Include/ScriptCanvas/Components/EditorScriptCanvasComponent.h
+    Editor/Include/ScriptCanvas/Components/EditorDeprecationData.h
+    Editor/Include/ScriptCanvas/Components/EditorDeprecationData.cpp
     Editor/Components/EditorScriptCanvasComponent.cpp
     Editor/Include/ScriptCanvas/Components/EditorScriptCanvasComponentSerializer.h
     Editor/Include/ScriptCanvas/Components/EditorScriptCanvasComponentSerializer.cpp
